Fix Reduce wastes I/Os on creating empty Reduction Rule 1 streams
This is the simple fix for the problem identified by @Tyilo . He suggested we
create a lazy_file_stream wrapping class of the tpie::file_stream to keep the
algorithm code cleaner. This is the smaller and simpler fix.

When we have to implement the `stable_reduce`, then we may want to reconsider
the above.