Use tpie::file_stream<T> for writing
An experiment shows that the `tpie::file<T>::stream` has a 6x performance loss
on writing to a file. So, usint `tpie::file_stream<T>` for writing and
`tpie::file<T>::stream` for parallel reading, then we achieve everything we
wanted from the refactor.

This requires the programmer to _detach_ the writer before one reads, calls a
bdd function, or merely reads off the size. For most reasonable use-cases this
does apply nicely.