Fix Tic-Tac-Toe example failing by cleaning up code
Why exactly something breaks by calling a function from a function, that
itself we never call anywhere else, I cannot explain. But, by not having
any clash in naming of the tuple structs, there seems to be no confusion
as to which kind of tuple to use.

I suppose the Type-checker and other analysis used the intended tuple
struct, yet the linker messed things up. Now, everything is cleaner, but
also much more importantly... it works.