Remove bdd_unique operator
The quantification algorithm (as is) tries to abuse, that any request (v,v) is
equivalent to (v,_), so it can be merged with the requests that survived through
the 'label' level. Yet, for the XOR this is not the case; in fact (v,v) will
always result in a false sink.

Similarly, the result of any node (v,_) (above or below the 'label' level) should
be the false sink, since there is no 'unique' assignment to the variable to change
the outcome.

Both things above, while interesting, do add unneeded computation to identify these
cases compared to what would be needed for the 'bdd_forall' and 'bdd_exists' cases.
Whether these added computations truly will be something one can see I have not
benchmarked, but until someone requests the 'bdd_unique' operation, I will keep
the code simplicity of 'bdd_forall' and 'bdd_exists'.