Make Isomorphism check provably O(sort(N))
We add an early termination when the number of unique requests to the
same level exceed the number of nodes on the same level in the original
two BDDs. If this is the case, then they cannot be isomorphic.

This also gives an 2N upper bound on the number of elements placed in the
priority queues and hence results in an O(sort(N)) I/O bound rather than
the prior O(sort(N^2))