Add safe to use 'memory_fits' function
The 'memory_usage' function can easily overflow, since we are creating a
larger number (especially with the levelized priority queue that doubles
and adds such numbers together).

This is a slightly hacky under-approximation, since a tpie::internal_vector
can include 2 more elements in a set amount of memory than can be placed
in the tpie::internal_priority_queue. But, I don't believe those two
elements are going to be worth the hassle.