Create access mode for prod2 and setup logic
Access mode controls whether to use the priority queues (old setup) or
random access to construct output layers. The logic checks which of
these access modes to used (if set to AUTO), and calls the appropiate
function.

The __prod2_ra functions is undefined.

All old bdd tests are setup to run the priority queue function.
zdd tests are unaltered.