Quick fix to compile Nested Sweeping on Clang-11
Clang-11 seems to have a slightly buggy overload resolution where the file type
without an implicit conversion constructor is on-par with the conversion.

I considered whether to just bump the minimum Clang version to 12 in the CI, but
for now this little change only adds a quite benign O(1) slowdown to each inner
sweep.