Use Non-canonical Reduce within Nested Sweeping (compile-time)
This adds usage of the 'nested_sweeping::__reduce_level__fast' algorithm within
'nested_sweep<>' based on a compile-time enum chosen. To this end, the interface
to the template parameters of 'nested_sweep' also has been cleaned up to only be
a single policy (from which the Reduce sweeps are derived).