Replace 'bdd_satmin' and 'bdd_satmax' implementation to use an internal memory stack
There is no need to make an external memory file for this algorithm. Adiar has a
minimum requirement of 128 MiB of memory, and even if all levels are present,
then they would only require up to 8 MiB of memory. This leaves enough space to
open all relevant streams and it will be considerably faster (since we do no
unecessary read/writes of tiny files to the disk)