Add .clang-format
After a bunch of testing, this Clang Format configuration is quite close
to what is done _most of the time_ in this repository. The only thing
I've found which it breaks consistently is when an UDT is instanciated
inline, the braces are split:
```
struct bar
{} foo;
```
becomes
```
struct bar
{
} foo;
```
and (maybe worse cause that's how it's often written in this project)
```
struct bar
{}
// comment
foo;
```
becomes
```
struct bar
{
}
// comment
foo;
```
. This along with disallowing the rest of multiple operator seperated
items in argument position to flow to the last line if they are
neccesary to split giving:
```
throw invalid_argument("Adiar requires at least "
                       + std::to_string(minimum_memory / 1024 / 1024) + " MiB of memory");
```
instead of
```
throw invalid_argument("Adiar requires at least "
                       + std::to_string(minimum_memory / 1024 / 1024)
		       + " MiB of memory");
```

Except for that, all changes I've observed it making is within
"normal" for the codebase.