Remove 'Terminal Count' heuristic
It seems like this heuristic is not good enough at differentiating between cases
that should use one or more partial sweeps and those that should not. If anything,
it might only be a bad proxy for some other meta information that we should use
instead. Hence, we better remove it to not have semi-dead and untested code laying
around in the codebase